Fix wrong layout length for longs

+++ src/main/java/de/uni_koblenz/west/koral/master/statisticsDB/impl/multi_file/log/CompressedLogReader.java
@@ -145,7 +145,7 @@ public class CompressedLogReader {
 				cursor += Short.BYTES;
 			} else if (elementType == ElementType.LONG) {
 				data.put(elementName, NumberConversion.bytes2long(row, cursor));
-				cursor += Short.BYTES;
+				cursor += Long.BYTES;
 			} else {
 				throw new IllegalArgumentException("Unkown element type: " + elementType);
 			}
